HV-1292 Add support for spec defined constraint @Email
